pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gtn-benchmark"
version = "0.1.0"
description = "GTN tutorial loader and Jetstream-powered benchmark generator"
authors = [{name = "Galaxy Benchmark"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "PyYAML>=6.0",
    "requests>=2.31",
    "tqdm>=4.66",
    "beautifulsoup4>=4.12"
]

[project.scripts]
generate-gtn-benchmark = "scripts.generate_benchmark:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["gtn_benchmark*"]
